# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Todo": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"done": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"done"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Todo.description-or-done-or-title_": {
				"properties": {
					"description": {
						"type": "string"
					},
					"done": {
						"type": "boolean"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"done",
					"title"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CreateTodoParameters": {
				"$ref": "#/components/schemas/Pick_Todo.description-or-done-or-title_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "tsoa-jest",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/todos": {
			"get": {
				"operationId": "GetTodos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Todo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTodo",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTodoParameters"
							}
						}
					}
				}
			}
		},
		"/api/v1/todos/{todoId}": {
			"put": {
				"operationId": "UpdateTodo",
				"responses": {
					"204": {
						"description": "Updated"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "todoId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Todo"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}